Add internal options.
Change the previous `optShow` boolean field in `Option` to an
`optVisibility` field which can be set to Visible, Hidden, or Internal.

Internal means that the option will not be shown in the help text at
all.

Add a new `internal` modifier to create internal options, and rename the
`hide` modifier to `hidden` for consistency.