Remove propDefault field and simplify builders.
A significant simplification of the `Option` type. Since the
introduction of the `Alternative` instance, there is no need for an
explicit default value anymore.

Also, builders can be simplified to work on a single type, as there's
rarely a need for polymorphic updates. As a consequence, the `transform`
builder has been removed.

To make an option with a default value of `Nothing`, now we can do:

    optional :: Parser a -> Parser (Maybe a)
    optional p = (Just <$> p) <|> pure Nothing

which is more general (it works on parsers, not just options), less
ad-hoc, and probably easier to understand and reason about.

The modifier parameter for `arguments` has now been changed to operate
on the final parser, instead of the underlying parser (fixes #14).