Automatically word wrap descriptions.
Now all the fields that might have to be formatted have type `Chunk
Doc`, and their builders create them using `paragraph`, so that they can
be automatically word-wrapped.

I also added alternative builders that take a `Maybe Doc` instead of a
string, which might be useful for descriptions that require special
formatting.