Add quick installation and import instructions
When I started trying to use this library I had to go look up which module to import on Hackage, so I thought it would be nice to have it at the start of the getting started tutorial. It might also be nice to say how to install it, so people don't have to know the package name.