Allow more complex usage of many and some.
Previously the entire BindP was required to pass during each call to
searchParser. Meaning that many pairs of arguments were not possible. By
being a bit more lazy in progressing through these options, and rebuilding
the BindP instead of asserting that it must be complete, we can allow for
many pairs or triples of items.

For example:
p = many $ (,) <$> argument str idm <*> argument str idm