Change how subparsers are executed:
Previously seachparser would return a Maybe a, which would then be
wrapped in a NilP . Just and swapped into the parser. This means
that subparsers needed to execute completely before returning, and
gives a somewhat unsatifactory notion of backtracking, where one
could supply commands to the top level, but only after fully
completing the subparser.

If instead of returning a Maybe a, we return a Maybe (Parse a) then
for flags, args, and options we can bahave the same and just wrap a
little earlier, but for subcommands we can do something a bit different
and replace the subparser with the contents of the selected command.

This means that we're essentially inlining the selected subparser,
and all parent options are still 'in scope' of the current parse.

Given that we can still enter the context (but not exit it), not much
changes with this, but backtracking is now quite different and can not
be turned off.