Add documentation on 'noBacktrack' preference
Subcommand parsers are often disjoint, meaning that if one fails
to parse after the command name then all others (having distinct
command names) will also fail to parse. When the 'noBacktrack'
preference is not set but 'showHelpOnError' is this leads to
the help for the top-level parser being shown instead of the
help for the failing subcommand.

A note has been added to the Haddock docs for the 'noBacktrack'
option to make this clear.