Bump pretty printer and stop relying on its internal modules.
This means that the prettyprinter library can now be used instead
of the ansi-wl-pprint library depending on the version bounds.

This was a bit tricky to get right, as there were a few tricks
we pulled using the internal modules. Figured it out though.

I've turned off deprection warnings on the module, which is a
bit rough, but otherwise it spews many, many warnings.