Handle parent groups without top-level options
It is possible that someone defines a "parent group" that only contains
child group(s) i.e. it does not contain any top-level options. This is
a problem, because the parent name will not be attached to any
options, thus will not be printed.

To fix this, we instead represent the Option Groups as a hierarchy
list of group titles e.g. ["Parent 1", "Parent 2", "Child group"].

Then, when rendering, we keep track of which group titles have already
been printed. If some group has parent titles that have not yet been
printed, we prepend their titles to the normal output.