Add maybeValue combinator.
This combinator is a convenience for handling default values that may
fail. I use this pattern to determine sane defaults from the system if
available but if not enforce the value to be required instead of having
a default.

I'm not sure if this is the best location nor what the recommended
testing strategy is for something like this, but please let me know if
there is interest in adding a new combinator such as this and what I can
improve to ensure it's successfully merged.